Split emit direct library validation tests
